Problem printing "quoteright"
Issue #42. If the character name is not in the post table, we continue
searching it in the cmaps.

This modification corrects the issue with the "quoteright", but then
some glyphs like "Ccaron" are no longer displayed.
If we remove the "differences" for WinAnsiEncoding in
PDFFontEncoding.java, they are displayed again.