[PDF_RENDERER-145] FIX conditionally required font dictionary properties
For Type 3 fonts, some font descriptor properties are optional that are required for all other font types. This should be taken into account when parsing a font descriptor.

For more details, see PDF Reference, Sixth Edition, version 1.7, section 5.7.

Additionally, we encountered a PDF where one of these parameters is also missing for a TrueType font. Therefore, a system property for lenient font descriptor parsing was introduced to allow rendering of this PDF as well.